+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file. For commit 
 
 ## v1.2.3-beta
 
+ **Notes**:
+ - Removed deprecated behavior for file uploads that checked for path ending in "/" to determine if is directory
+ - renamed public upload api query param from `targetPath` to `path`. see swagger docs.
 
  **BugFixes**:
  - fix stuck motion preview popup window issue
+ - Chunked uploads ignore User Scope and are saved to Source Root (#1894)
+ - share delete with user scope issue
 
 ## v1.2.2-beta